requirements.txt
langgraph
langchain-openai
langchain-community
langchain-text-splitters
langchain-core
python-dotenv
faiss-cpu